[refactor] API 응답 에러 메세지 분리
